backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.52.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "mcp>=1.9.1",
    "openai>=1.82.0",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "supabase>=2.15.1",
]
